requirement.txt
Django==2.2.7
django-crispy-forms==1.8.0
django-widget-tweaks==1.4.5
djangorestframework==3.10.3
mysqlclient==1.4.5
PyMySQL==0.9.3
pytz==2019.3
sqlparse==0.3.0
